# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Almuzini Children Hospital EHR API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /prescriptions:
    get:
      operationId: listPrescriptions
      tags: [prescriptions]
      summary: List prescriptions
      parameters:
        - name: patientId
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Prescriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prescription"
    post:
      operationId: createPrescription
      tags: [prescriptions]
      summary: Create prescription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrescriptionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
  /prescriptions/{id}:
    get:
      operationId: getPrescription
      tags: [prescriptions]
      summary: Get prescription
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Prescription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
    patch:
      operationId: updatePrescription
      tags: [prescriptions]
      summary: Update prescription status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrescriptionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"

  # LAB ORDERS
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List staff members
      parameters:
        - name: role
          in: query
          schema: { type: string }
        - name: department
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
    post:
      operationId: createStaff
      tags: [staff]
      summary: Create staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
  /staff/{id}:
    get:
      operationId: getStaffMember
      tags: [staff]
      summary: Get staff member
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
    patch:
      operationId: updateStaff
      tags: [staff]
      summary: Update staff member
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"

  # VACCINATIONS
  /vaccinations:
    get:
      operationId: listVaccinations
      tags: [vaccinations]
      summary: List vaccinations for a patient
      parameters:
        - name: patientId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Vaccinations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vaccination"
    post:
      operationId: recordVaccination
      tags: [vaccinations]
      summary: Record vaccination
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaccinationInput"
      responses:
        "201":
          description: Recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vaccination"

  # GROWTH CHARTS
  /growth-records:
    get:
      operationId: listGrowthRecords
      tags: [growth]
      summary: List growth records for a patient
      parameters:
        - name: patientId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Growth records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GrowthRecord"
    post:
      operationId: createGrowthRecord
      tags: [growth]
      summary: Record growth measurement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrowthRecordInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GrowthRecord"

  # DASHBOARD / WOW ENDPOINTS
...
components:
  schemas:
...
    Prescription:
      type: object
      required: [id, patientId, prescriberId, drugName, dosage, frequency, status, createdAt]
      properties:
        id: { type: integer }
        patientId: { type: integer }
        patientName: { type: ["string", "null"] }
        prescriberId: { type: integer }
        prescriberName: { type: ["string", "null"] }
        drugName: { type: string }
        drugNameAr: { type: ["string", "null"] }
        dosage: { type: string }
        frequency: { type: string }
        duration: { type: ["string", "null"] }
        route: { type: ["string", "null"] }
        instructions: { type: ["string", "null"] }
        status: { type: string }
        createdAt: { type: string }
    PrescriptionInput:
      type: object
      required: [patientId, drugName, dosage, frequency]
      properties:
        patientId: { type: integer }
        drugName: { type: string }
        drugNameAr: { type: string }
        dosage: { type: string }
        frequency: { type: string }
        duration: { type: string }
        route: { type: string }
        instructions: { type: string }
    PrescriptionUpdate:
      type: object
      properties:
        status: { type: string }
        instructions: { type: string }

    # LAB ORDERS
...
    StaffMember:
      type: object
      required: [id, nameEn, role, department, createdAt]
      properties:
        id: { type: integer }
        nameEn: { type: string }
        nameAr: { type: ["string", "null"] }
        role: { type: string }
        department: { type: string }
        specialization: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        licenseNumber: { type: ["string", "null"] }
        status: { type: string }
        createdAt: { type: string }
    StaffInput:
      type: object
      required: [nameEn, role, department]
      properties:
        nameEn: { type: string }
        nameAr: { type: string }
        role: { type: string }
        department: { type: string }
        specialization: { type: string }
        email: { type: string }
        phone: { type: string }
        licenseNumber: { type: string }
    StaffUpdate:
      type: object
      properties:
        nameEn: { type: string }
        nameAr: { type: string }
        department: { type: string }
        specialization: { type: string }
        phone: { type: string }
        status: { type: string }

    # VACCINATIONS
    Vaccination:
      type: object
      required: [id, patientId, vaccineName, administeredDate, createdAt]
      properties:
        id: { type: integer }
        patientId: { type: integer }
        vaccineName: { type: string }
        vaccineNameAr: { type: ["string", "null"] }
        doseNumber: { type: ["integer", "null"] }
        administeredDate: { type: string }
        nextDueDate: { type: ["string", "null"] }
        batchNumber: { type: ["string", "null"] }
        administeredById: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    VaccinationInput:
      type: object
      required: [patientId, vaccineName, administeredDate]
      properties:
        patientId: { type: integer }
        vaccineName: { type: string }
        vaccineNameAr: { type: string }
        doseNumber: { type: integer }
        administeredDate: { type: string }
        nextDueDate: { type: string }
        batchNumber: { type: string }
        notes: { type: string }

    # GROWTH RECORDS
    GrowthRecord:
      type: object
      required: [id, patientId, measurementDate, createdAt]
      properties:
        id: { type: integer }
        patientId: { type: integer }
        measurementDate: { type: string }
        weight: { type: ["number", "null"] }
        height: { type: ["number", "null"] }
        headCircumference: { type: ["number", "null"] }
        bmi: { type: ["number", "null"] }
        weightPercentile: { type: ["number", "null"] }
        heightPercentile: { type: ["number", "null"] }
        bmiPercentile: { type: ["number", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    GrowthRecordInput:
      type: object
      required: [patientId, measurementDate]
      properties:
        patientId: { type: integer }
        measurementDate: { type: string }
        weight: { type: number }
        height: { type: number }
        headCircumference: { type: number }
        notes: { type: string }

    # DASHBOARD
...