# lib/api-zod/src/generated/types/staffInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Almuzini Children Hospital EHR API
 * OpenAPI spec version: 0.1.0
 */

export interface StaffInput {
  nameEn: string;
  nameAr?: string;
  role: string;
  department: string;
  specialization?: string;
  email?: string;
  phone?: string;
  licenseNumber?: string;
}